requirements.txt
shortuuid
python-dotenv
requests